requirements.txt
astroid==2.4.2
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.3
Flask==1.1.2
Flask-SQLAlchemy==2.4.3
idna==2.9
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
pycurl==7.43.0
pygobject==3.20.0
pylint==2.5.3
python-apt==1.1.0b1+ubuntu0.16.4.8
python-dotenv==0.13.0
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.17
tmdbsimple==2.3.0
toml==0.10.1
typed-ast==1.4.1
unattended-upgrades==0.1
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.12.1
